+++ source/Private/Invoke-GuestConfigurationPackage.ps1
@@ -18,14 +18,35 @@ function Invoke-GuestConfigurationPackage
         $Apply
     )
 
-    if ($IsMacOS)
+    $os = Get-OSPlatform
+    if ($os -ieq 'MacOS')
     {
-        throw 'The Invoke-GuestConfigurationPackage cmdlet is not supported on MacOS'
+        throw 'This cmdlet is not supported on MacOS.'
     }
 
-    $Path = Resolve-RelativePath -Path $Path
+    #-----VALIDATE PARAMETERS-----
+    $requiredParameterProperties = @('ResourceType', 'ResourceId', 'ResourcePropertyName', 'ResourcePropertyValue')
+
+    foreach ($parameterInfo in $Parameter)
+    {
+        foreach ($requiredParameterProperty in $requiredParameterProperties)
+        {
+            if (-not ($parameterInfo.Keys -contains $requiredParameterProperty))
+            {
+                $requiredParameterPropertyString = $requiredParameterProperties -join ', '
+                throw "One of the specified parameters is missing the mandatory property '$requiredParameterProperty'. The mandatory properties for parameters are: $requiredParameterPropertyString"
+            }
+
+            if ($parameterInfo[$requiredParameterProperty] -isnot [string])
+            {
+                $requiredParameterPropertyString = $requiredParameterProperties -join ', '
+                throw "The property '$requiredParameterProperty' of one of the specified parameters is not a string. All parameter property values must be strings."
+            }
+        }
+    }
 
     #-----VALIDATE PACKAGE SETUP-----
+    $Path = Resolve-RelativePath -Path $Path
 
     if (-not (Test-Path -Path $Path -PathType 'Leaf'))
     {