+++ source/Public/Test-GuestConfigurationPackage.ps1
@@ -86,11 +86,6 @@ function Test-GuestConfigurationPackage
         $Parameter = @()
     )
 
-    if ($IsMacOS)
-    {
-        throw 'The Test-GuestConfigurationPackage cmdlet is not supported on MacOS'
-    }
-
     $invokeParameters = @{
         Path = $Path
     }